+++ src/constants/index.ts
@@ -1,0 +1,2 @@
+
+export const onPropRegex = /on[A-Z](.*)/;